+++ js/packages/phoenix-cli/src/commands/index.ts
@@ -1,7 +1,7 @@
-// Command exports
-// This file will export all available CLI commands
-
-// TODO: Add command exports here
-// export * from "./prompts";
-// export * from "./datasets";
-// export * from "./experiments";
+export * from "./projects";
+export * from "./traces";
+export * from "./trace";
+export * from "./datasets";
+export * from "./dataset";
+export * from "./experiments";
+export * from "./experiment";