+++ js/packages/phoenix-evals/src/llm/index.ts
@@ -1,8 +1,9 @@
-export * from "./generateClassification";
-export * from "./createClassifierFn";
-export * from "./createClassificationEvaluator";
-export * from "./createHallucinationEvaluator";
-export * from "./createDocumentRelevanceEvaluator";
-export * from "./createCorrectnessEvaluator";
 export * from "./ClassificationEvaluator";
+export * from "./createClassificationEvaluator";
+export * from "./createClassifierFn";
+export * from "./createCorrectnessEvaluator";
+export * from "./createDocumentRelevanceEvaluator";
+export * from "./createFaithfulnessEvaluator";
+export * from "./createHallucinationEvaluator"; // Deprecated: use createFaithfulnessEvaluator
+export * from "./generateClassification";
 export * from "./LLMEvaluator";