+++ js/examples/apps/demo-document-relevancy-experiment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # demo-document-relevancy-experiment
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [01eb1fb]
+  - @arizeai/phoenix-client@5.7.0
+
 ## 1.0.6
 
 ### Patch Changes